--- app/helpers/ui/color-scheme-init.entry.ts
@@ -1,3 +1,0 @@
-import { initColorScheme } from "./color-scheme"
-
-initColorScheme()
