+++ app/helpers/ui/colorSchemaInit.entry.ts
@@ -1,0 +1,3 @@
+import { initColorScheme } from "./colorSchema"
+
+initColorScheme()
